+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * 【core 】     新增WatchServer（issue#440@Github）
 * 【core 】     ReflectUtil.getFieldValue支持static（issue#662@Github）
 * 【core 】     改进Bean判断和注入逻辑：支持public字段注入（issue#I1689L@Gitee）
+* 【extra】     新增SpringUtil
 
 
 ### Bug修复